Implement pre-encoded namespace diffs
