+++ share-utils/src/Share/Utils/Aeson.hs
@@ -6,12 +6,11 @@ import Data.Aeson (ToJSON (..))
 import Data.Aeson qualified as Aeson
 import Data.Aeson.Encoding qualified as Encoding
 import Data.Binary.Builder qualified as Builder
-import Data.Text.Lazy qualified as TL
-import Data.Text.Lazy.Encoding qualified as TL
+import Data.ByteString.Lazy.Char8 qualified as BL
 import GHC.Stack (HasCallStack)
 
 data MaybeEncoded a
-  = IsEncoded TL.Text
+  = IsEncoded BL.ByteString
   | NotEncoded a
   deriving stock (Show, Eq, Ord, Functor, Foldable, Traversable)
 
@@ -24,12 +23,12 @@ instance (ToJSON a) => ToJSON (MaybeEncoded a) where
   toEncoding (IsEncoded txt) = Aeson.toEncoding (PreEncoded txt)
   toEncoding (NotEncoded a) = Aeson.toEncoding a
 
-newtype PreEncoded = PreEncoded TL.Text
+newtype PreEncoded a = PreEncoded BL.ByteString
   deriving stock (Show, Eq, Ord)
 
-instance ToJSON PreEncoded where
-  toJSON :: (HasCallStack) => PreEncoded -> Aeson.Value
+instance ToJSON (PreEncoded a) where
+  toJSON :: (HasCallStack) => PreEncoded a -> Aeson.Value
   toJSON (PreEncoded _txt) = error "Tried to encode a PreEncoded value back into a Value. Ensure you're using toEncoding instead."
 
-  toEncoding :: (HasCallStack) => PreEncoded -> Aeson.Encoding
-  toEncoding (PreEncoded txt) = Encoding.unsafeToEncoding $ Builder.fromLazyByteString $ TL.encodeUtf8 txt
+  toEncoding :: (HasCallStack) => PreEncoded a -> Aeson.Encoding
+  toEncoding (PreEncoded txt) = Encoding.unsafeToEncoding . Builder.fromLazyByteString $ txt
